Extract auth client to a base service object
