+++ lib/g5_authenticatable_api/base_service.rb
@@ -1,0 +1,16 @@
+module G5AuthenticatableApi
+  class BaseService
+    def access_token
+      raise 'Not implemented'
+    end
+
+    def token_info
+      auth_client.token_info
+    end
+
+    def auth_client
+      @auth_client ||= G5AuthenticationClient::Client.new(allow_password_credentials: 'false',
+                                                          access_token: access_token)
+    end 
+  end
+end
